Update cert inheritance tests to match new xapi behavior
For XCP-ng > 8.2.1 there is a new behavior regarding SB certs management
See: https://github.com/xapi-project/xen-api/pull/4659

Keep previous behavior for XCP-ng <= 8.2.1
